Move sliders as Front page subcomponents
