+++ src/Front/Homes.js
@@ -2,9 +2,9 @@ import React from "react";
 
 import bp from "../breakpoints";
 import { Slider, Slide } from "../UI/Slider";
-import Card from "./Card";
+import Card from "../Homes/Card";
 
-import data from "./staticData";
+import data from "../Homes/staticData";
 
 const HomeSlide = Slide.extend`
   max-width: 197px;
