+++ src/Front/Reservations.js
@@ -1,9 +1,9 @@
 import React from "react";
 
 import { Slider } from "../UI/Slider";
-import Card from "./Card";
+import Card from "../Restaurants/Card";
 
-import data from "./staticData";
+import data from "../Restaurants/staticData";
 
 export default function() {
   const restaurantsList = data.map(restaurant => (
